% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-models.R
\docType{data}
\name{models}
\alias{models}
\title{Pre-made models usable by the decryptr package}
\format{
A tibble with 4 rows and 2 variables:
\describe{
\item{name}{Name of the model}
\item{labs}{All possible lables the model could output}
}
}
\usage{
models
}
\description{
Pre-made models usable by the decryptr package
}
\keyword{datasets}
