% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-model.R
\name{read_model}
\alias{read_model}
\title{Read keras model by regex}
\usage{
read_model(model_name)
}
\arguments{
\item{model_name}{regex to find model inside \code{system.file}.}
}
\description{
Read keras model by regex
}
\examples{

\dontrun{
read_model('mg')
}

}
